+++ com.unity.multiplayer.mlapi/Runtime/Metrics/INetworkMetrics.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01a54e8729c7e3a4bb566d9274c44cb3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
