+++ com.unity.multiplayer.mlapi/Runtime/Metrics/NullNetworkMetrics.cs
@@ -1,0 +1,39 @@
+﻿using System.Collections.Generic;
+
+namespace MLAPI.Metrics
+{
+    public class NullNetworkMetrics : INetworkMetrics
+    {
+        public void TrackNetworkObject(NetworkObject networkObject)
+        {
+        }
+
+        public void TrackNamedMessageSent(ulong receiverClientId, string messageName, ulong bytesCount)
+        {
+        }
+
+        public void TrackNamedMessageSent(IReadOnlyCollection<ulong> receiverClientIds, string messageName, ulong bytesCount)
+        {
+        }
+
+        public void TrackNamedMessageReceived(ulong senderClientId, string messageName, ulong bytesCount)
+        {
+        }
+
+        public void TrackUnnamedMessageSent(ulong receiverClientId, ulong bytesCount)
+        {
+        }
+
+        public void TrackUnnamedMessageSent(IReadOnlyCollection<ulong> receiverClientIds, ulong bytesCount)
+        {
+        }
+
+        public void TrackUnnamedMessageReceived(ulong senderClientId, ulong bytesCount)
+        {
+        }
+
+        public void DispatchFrame()
+        {
+        }
+    }
+}
