+++ com.unity.multiplayer.mlapi/Runtime/Metrics/NullNetworkMetrics.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b858c4c3c7e08884bafdba745e688dbb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
